Apply changes suggested in code review.
