+++ resources/service-catalog/templates/tests/rbac.yaml
@@ -24,19 +24,15 @@ rules:
   - apiGroups: ["servicecatalog.k8s.io"]
     resources: ["clusterserviceclasses"]
     verbs: ["list", "delete", "get"]
-    nonResourceURLs: []
   - apiGroups: ["servicecatalog.k8s.io"]
     resources: ["clusterservicebrokers"]
     verbs: ["get", "list"]
-    nonResourceURLs: []
   - apiGroups: ["servicecatalog.k8s.io"]
     resources: ["clusterserviceplans"]
     verbs: ["list", "delete"]
-    nonResourceURLs: []
   - apiGroups: ["servicecatalog.k8s.io"]
     resources: ["serviceinstances", "servicebindings"]
     verbs: ["get", "delete", "create", "list"]
-    nonResourceURLs: []
   - apiGroups: ["servicecatalog.k8s.io"]
     resources: ["servicebrokers"]
     verbs: ["get", "list"]
@@ -46,35 +42,27 @@ rules:
   - apiGroups: ["servicecatalog.kyma-project.io"]
     resources: ["servicebindingusages"]
     verbs: ["get", "delete", "create", "list"]
-    nonResourceURLs: []
   - apiGroups: ["applicationconnector.kyma-project.io"]
     resources: ["applications", "applicationmappings"]
     verbs: ["get", "list", "create", "delete"]
-    nonResourceURLs: []
   - apiGroups: ["apps"]
     resources: ["deployments"]
     verbs: ["create", "delete", "get", "list"]
-    nonResourceURLs: []
   - apiGroups: ["addons.kyma-project.io"]
     resources: ["addonsconfigurations", "clusteraddonsconfigurations"]
     verbs: ["create", "delete", "get", "list"]
-    nonResourceURLs: []
   - apiGroups: [""]
     resources: ["services", "namespaces", "pods", "pods/log", "events", "configmaps", "serviceaccounts"]
     verbs: ["create", "delete", "get", "list", "update"]
-    nonResourceURLs: []
   - apiGroups: [""]
     resources: ["pods/exec"]
     verbs: ["create"]
-    nonResourceURLs: []
   - apiGroups: ["config.istio.io"]
     resources: ["rules", "checknothings", "deniers"]
     verbs: ["create", "delete", "get", "list"]
-    nonResourceURLs: []
   - apiGroups: ["rbac.authorization.k8s.io"]
     resources: ["roles", "rolebindings"]
     verbs: ["create"]
-    nonResourceURLs: []
 ---
 kind: ClusterRoleBinding
 apiVersion: rbac.authorization.k8s.io/v1beta1
